pyproject.toml
[project]
name = "kursovaya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.6.0",
    "SQLAlchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "python-dotenv>=1.0.0",
    "sympy>=1.12",
    "reportlab>=4.0.0",
    "qrcode[pil]>=7.4",
    "pytest>=7.4.0",
]
